+++ src/components/com_kunena/controller/topic/moderate/display.php
@@ -117,7 +117,7 @@ class ComponentKunenaControllerTopicModerateDisplay extends KunenaControllerDisp
 
 			if ($user->exists())
 			{
-			    $username       = KunenaFactory::getProfile()->getProfileName($user);
+				$username       = $user->getName();
 				$this->userLink = $this->message->userid ? HTMLHelper::_('kunenaforum.link',
 					'index.php?option=com_kunena&view=user&userid=' . $this->message->userid,
 					$username . ' (' . $this->message->userid . ')', $username . ' (' . $this->message->userid . ')'