+++ src/components/com_kunena/template/crypsis/layouts/user/item/default.php
@@ -16,7 +16,7 @@ $tabs = $this->getTabs();
 
 <h1 class="pull-left">
 	<?php echo Text::_('COM_KUNENA_USER_PROFILE'); ?>
-	<?php echo $this->escape(KunenaFactory::getProfile()->getProfileName($this->profile)); ?>
+	<?php echo $this->escape($this->profile->getName()); ?>
 </h1>
 
 <h2 class="pull-right">