+++ src/components/com_kunena/template/crypsisb3/layouts/message/row/default.php
@@ -94,7 +94,7 @@ $txt             = $topic->getActions();
 				?>
 				<?php echo $topic->getFirstPostTime()->toKunena('config_post_dateformat'); ?>
 				<?php echo Text::_('COM_KUNENA_BY') ?>
-				<?php echo $topic->getAuthor()->getLink(null, Text::sprintf('COM_KUNENA_VIEW_USER_LINK_TITLE', KunenaFactory::getProfile()->getProfileName($topic->getAuthor())), '', '', KunenaTemplate::getInstance()->tooltips(), $category->id); ?>
+				<?php echo $topic->getAuthor()->getLink(null, Text::sprintf('COM_KUNENA_VIEW_USER_LINK_TITLE', $topic->getAuthor()->getName()), '', '', KunenaTemplate::getInstance()->tooltips(), $category->id); ?>
 			<?php endif; ?>
 			<div class="pull-right">
 				<?php /** TODO: New Feature - LABELS
@@ -124,12 +124,12 @@ $txt             = $topic->getActions();
 			<div class="row-fluid">
 				<?php if ($config->avataroncat) : ?>
 					<div class="col-md-3">
-						<?php echo $topic->getLastPostAuthor()->getLink($avatar, Text::sprintf('COM_KUNENA_VIEW_USER_LINK_TITLE', KunenaFactory::getProfile()->getProfileName($topic->getLastPostAuthor())), '', '', KunenaTemplate::getInstance()->tooltips(), $category->id); ?>
+						<?php echo $topic->getLastPostAuthor()->getLink($avatar, Text::sprintf('COM_KUNENA_VIEW_USER_LINK_TITLE', $topic->getLastPostAuthor()->getName()), '', '', KunenaTemplate::getInstance()->tooltips(), $category->id); ?>
 					</div>
 				<?php endif; ?>
 				<div class="col-md-9">
 					<span class="datepost"><?php echo $this->getTopicLink($topic, 'last', Text::_('COM_KUNENA_GEN_LAST_POST'), null, KunenaTemplate::getInstance()->tooltips(), $category, false, true); ?>
-						<?php echo ' ' . Text::_('COM_KUNENA_BY') . ' ' . $topic->getLastPostAuthor()->getLink(null, Text::sprintf('COM_KUNENA_VIEW_USER_LINK_TITLE', KunenaFactory::getProfile()->getProfileName($topic->getLastPostAuthor())), '', '', KunenaTemplate::getInstance()->tooltips(), $category->id); ?>
+						<?php echo ' ' . Text::_('COM_KUNENA_BY') . ' ' . $topic->getLastPostAuthor()->getLink(null, Text::sprintf('COM_KUNENA_VIEW_USER_LINK_TITLE', $topic->getLastPostAuthor()->getName()), '', '', KunenaTemplate::getInstance()->tooltips(), $category->id); ?>
 					</span>
 					<br/>
 					<span class="datepost"><?php echo $topic->getLastPostTime()->toKunena('config_post_dateformat'); ?></span>